requirements.txt
# Core data science libraries
pandas==1.5.3
numpy==1.23.5
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.3.0

# Statistical tests
scipy==1.10.1

# CTGAN and table evaluation
ctgan==0.7.1
table-evaluator==1.3.1

# Imbalanced data handling
imbalanced-learn==0.11.0

# Reinforcement Learning
stable-baselines3==1.7.0
gym==0.26.2
gymnasium==0.29.1
pygame==2.5.2

# Computer Vision & Deep Learning
torch==2.1.0
opencv-python==4.8.0.76
Pillow==9.5.0
ultralytics==8.0.20
roboflow==0.2.27

# Visualization
ipython==8.12.2  # For displaying in notebooks, e.g. display()

# Optional (used in notebooks or interactive environments)
jupyter
